x86: add OPTION_STATIC_OPTION_TABLE_NAME
This option uses a function instead of the hard-coded "cmos.default" for
the static option table.
This allows, say, a jumper on the mainboard to select a CMOS default file
that contains a 'debug' log level.
Building and adding alternate cmos.default files is an exercise left to
the reader.

Change-Id: Ieb4d255c0e9dd1b5de4214a37635146fd840b52b
Signed-off-by: Ben Gardner <gardner.ben@gmail.com>